合并login.ts和register.ts
